// src/modules/storefront/interfaces/Theme.ts
export interface MainModuleThemeInterface {
  name: string;
  type: ModulesType;
  id: string;
  styleData?: any;
  contentData?: any;
  deviceType?: 'none' | 'mobile' | 'desktop';
  languageType?: string;
  mobileStyleData?: any;
  mobileContentData?: any;
}
// ...
export interface AssetInterface {
  assetId: string;
  assetUrl: string;
  basePath?: string;
}
// ...
export enum AlignmentEnum {
  LEFT = 'left',
  CENTER = 'center',
  RIGHT = 'right',
}
// ...
export interface MidiaData extends MainModuleThemeInterface {
  type: ModulesType.MIDIA;
  styleData: {
    margin?: string;
    padding?: string;
    mediaUrl: AssetInterface;
    imageCompression?: 'best' | 'good' | 'eco' | 'no-compression';
    mediaUrlMobile: AssetInterface;
    mediaLink: string;
    imageDisposition: string;
    imageRatio: string;
    imageAlignment: AlignmentEnum;
  };
  mobileStyleData: MidiaData['styleData'];
}
// ...
export enum ModulesType {
  HEADER = 'Header',
  CONFIGURATION = 'Configuration',
  CATEGORIES = 'Categories',
  BANNER = 'Banner',
  CARDS = 'Cards',
  FOOTER = 'Footer',
  COOKIE = 'Cookie',
  ACCORDIONS = 'Accordions',
  IMAGE_PLUS_TEXT = 'Imagem + Texto',
  PARAGRAPH = 'Paragraph',
  PRODUCT_PAGE = 'Product_page',
  MIDIA = 'Midia',
  GRID_ITEM_AREA = 'GridItemArea',
  DYNAMIC_API = 'DynamicApi',
  TABLE = 'Table',
  BANNER_VARIANT = 'BannerVariant',
  CONTENT_CARD = 'Content Card',
  BUTTON = 'Button',
  BANNER_WJJC = 'BannerWJJC',
  PASS_BENEFIT = 'Pass - Benefício',
  REDIRECT = 'Redirect',
}